backend/requirements.txt
fastapi
uvicorn
celery
redis
crewai
litellm==1.74.0
openai==1.83.0
lancedb
pydantic
httpx
sse-starlette
gevent
SQLAlchemy
python-dotenv
chromadb
cryptography
alembic
passlib[bcrypt]
stripe
python-jose[cryptography]
python-multipart
ruff
